Bug 4894: -k restart shuts the instance down instead, pt 2
* Do not restart the coordinator process. Doing this properly is tricky
  and left as a TODO.

* Protect "restarting" mode from other commands with a counter.
  The restarting counter drops to zero when all kids are restarted.
  To avoid "restarting" forever in case of kids, irresponsive to
  signals, an alarm is set with 10-seconds delay by default.

* Partially undone 0c7a559: since there is no any objective need in
  forgetting only hopeless failures on restarting, we should
  follow the reconfiguring approach and forget all failures.

+++ src/tools.cc
@@ -360,12 +360,14 @@ death(int sig)
 }
 
 void
-BroadcastSignalIfAny(int& sig)
+BroadcastSignalIfAny(int& sig, const bool includingCoordinator)
 {
     if (sig > 0) {
         if (IamMasterProcess()) {
             for (int i = TheKids.count() - 1; i >= 0; --i) {
                 const auto &kid = TheKids.get(i);
+                if (!includingCoordinator && kid.isCoordinator())
+                    continue;
                 if (kid.running())
                     kill(kid.getPid(), sig);
             }
